@@ -1,31 +1,39 @@
 {
-  "name": "developer-portfolio",
+  "name": "portfolio.github.io",
+  "homepage": "https://pankajsbisht143.github.io",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-fast-marquee": "^1.2.1",
-    "react-helmet": "^6.1.0",
-    "react-icons": "^4.2.0",
-    "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-router-hash-link": "^2.4.3",
-    "react-scripts": "^4.0.3",
-    "react-slick": "^0.28.1",
-    "slick-carousel": "^1.8.1",
-    "validator": "^13.6.0",
-    "web-vitals": "^1.0.1"
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-brands-svg-icons": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "aos": "^2.3.4",
+    "font-awesome": "^4.7.0",
+    "gh-pages": "^5.0.0",
+    "latest": "^0.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-github-calendar": "^4.0.0",
+    "react-icons": "^4.8.0",
+    "react-particles": "^2.10.1",
+    "react-scripts": "5.0.1",
+    "save-dev": "^0.0.1-security",
+    "tslib": "^2.6.0",
+    "tsparticles": "^2.10.1",
+    "tsparticles-engine": "^2.10.1",
+    "typewriter-effect": "^2.19.0",
+    "uuid": "^9.0.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
     "start": "react-scripts start",
-    "build": "react-scripts --openssl-legacy-provider build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
